+++ NEWS.md
@@ -16,6 +16,7 @@
 * Improved performance and better documentation for `corpus_segment()`. (#634)
 * Added functions `corpus_trimsentences()` and `char_trimsentences()` to remove sentences from a corpus or character object, based on token length or pattern matching.
 * Added options to `textstat_readability()`: `min_sentence_length` and `max_sentence_length`. (#632)
+* Indexing now works for dictionaries, for slicing out keys and values (`[`), or accessing values directly (`[[`).  (#651)
 
 ### Behaviour changes